// src/app/model/usuario-model.ts
export class usuarioModel{
    id_usuario: number = 0;
    fk_rol: number = 0;
    usuario: String = '';
    pass: String = '';
    nombre: String = '';
    telefono: String = '';
    correo: String = '';
    fecha_creacion: String = '';
    usuario_creacion: String = '';
    fecha_mod: String = '';
    usuario_mod: String = '';
    estado: boolean = false
}